+++ timings.txt
@@ -1,0 +1,283 @@
+{
+ "1,1": 0
+}{
+ "1,1": 0
+}{
+ "1,11": 0
+}{
+ "1,21": 0
+}{
+ "1,31": 0
+}{
+ "1,41": 0
+}{
+ "1,51": 0
+}{
+ "1,61": 0
+}{
+ "1,71": 0
+}{
+ "1,81": 0
+}{
+ "1,91": 0
+}{
+ "11,1": 0
+}{
+ "11,11": 0
+}{
+ "11,21": 0
+}{
+ "11,31": 0
+}{
+ "11,41": 0
+}{
+ "11,51": 0
+}{
+ "11,61": 0
+}{
+ "11,71": 0
+}{
+ "11,81": 0
+}{
+ "11,91": 0
+}{
+ "21,1": 0
+}{
+ "21,11": 0
+}{
+ "21,21": 0
+}{
+ "21,31": 0
+}{
+ "21,41": 0
+}{
+ "21,51": 0
+}{
+ "21,61": 0
+}{
+ "21,71": 0
+}{
+ "21,81": 0
+}{
+ "31,1": 0
+}{
+ "31,11": 0
+}{
+ "31,21": 0
+}{
+ "31,31": 0
+}{
+ "31,41": 0
+}{
+ "31,51": 0
+}{
+ "31,61": 0
+}{
+ "31,71": 0
+}{
+ "31,81": 1
+}{
+ "31,91": 1
+}{
+ "41,1": 0
+}{
+ "41,11": 0
+}{
+ "41,21": 0
+}{
+ "41,31": 0
+}{
+ "41,41": 0
+}{
+ "41,51": 0
+}{
+ "41,61": 0
+}{
+ "41,71": 1
+}{
+ "41,81": 1
+}{
+ "41,91": 1
+}{
+ "51,1": 0
+}{
+ "51,11": 0
+}{
+ "51,21": 0
+}{
+ "51,31": 0
+}{
+ "51,41": 0
+}{
+ "51,51": 0
+}{
+ "51,61": 1
+}{
+ "51,71": 1
+}{
+ "51,81": 1
+}{
+ "51,91": 1
+}{
+ "61,1": 0
+}{
+ "61,11": 0
+}{
+ "61,21": 0
+}{
+ "61,31": 0
+}{
+ "61,41": 0
+}{
+ "61,51": 1
+}{
+ "61,61": 1
+}{
+ "61,71": 1
+}{
+ "61,81": 1
+}{
+ "61,91": 1
+}{
+ "71,1": 0
+}{
+ "71,11": 0
+}{
+ "71,21": 0
+}{
+ "71,31": 0
+}{
+ "71,41": 1
+}{
+ "71,51": 1
+}{
+ "71,61": 1
+}{
+ "71,71": 1
+}{
+ "71,81": 1
+}{
+ "71,91": 1
+}{
+ "81,1": 0
+}{
+ "81,11": 0
+}{
+ "81,21": 0
+}{
+ "81,31": 1
+}{
+ "81,41": 1
+}{
+ "81,51": 1
+}{
+ "81,61": 1
+}{
+ "81,71": 1
+}{
+ "81,81": 1
+}{
+ "81,91": 1
+}{
+ "91,1": 0
+}{
+ "91,11": 0
+}{
+ "91,21": 1
+}{
+ "91,31": 1
+}{
+ "91,41": 1
+}{
+ "91,51": 1
+}{
+ "91,61": 1
+}{
+ "91,71": 1
+}{
+ "91,81": 1
+}{
+ "91,91": 1
+}{
+ "101,1": 0
+}{
+ "101,11": 1
+}{
+ "101,21": 1
+}{
+ "101,31": 1
+}{
+ "101,41": 1
+}{
+ "101,51": 1
+}{
+ "101,61": 1
+}{
+ "101,71": 1
+}{
+ "101,81": 1
+}{
+ "101,91": 1
+}{
+ "111,1": 1
+}{
+ "111,11": 1
+}{
+ "111,21": 1
+}{
+ "111,31": 1
+}{
+ "111,41": 1
+}{
+ "111,51": 1
+}{
+ "111,61": 1
+}{
+ "111,71": 1
+}{
+ "111,81": 1
+}{
+ "111,91": 1
+}{
+ "21,91": 1
+}{
+ "1,91": 0
+}{
+ "1,101": 0
+}{
+ "1,111": 1
+}{
+ "11,91": 0
+}{
+ "11,101": 1
+}{
+ "11,111": 1
+}{
+ "21,91": 1
+}{
+ "21,101": 1
+}{
+ "21,111": 1
+}{
+ "31,91": 1
+}{
+ "31,101": 1
+}{
+ "31,111": 1
+}{
+ "41,91": 1
+}{
+ "41,101": 1
+}{
+ "101,1": 0
+}{
+ "101,11": 1
+}{
+ "111,1": 1
+}{
+ "111,11": 1
+}{
+ "121,1": 1
+}{
+ "121,11": 1
+}
