+++ timings10.txt
@@ -1,0 +1,421 @@
+{
+ "10,10": 0
+}{
+ "10,20": 0
+}{
+ "10,30": 0
+}{
+ "10,40": 0
+}{
+ "10,50": 0
+}{
+ "10,60": 0
+}{
+ "10,70": 0
+}{
+ "10,80": 0
+}{
+ "10,90": 0
+}{
+ "10,100": 1
+}{
+ "10,110": 1
+}{
+ "10,120": 1
+}{
+ "10,130": 1
+}{
+ "10,140": 1
+}{
+ "20,10": 0
+}{
+ "20,20": 0
+}{
+ "20,30": 0
+}{
+ "20,40": 0
+}{
+ "20,50": 0
+}{
+ "20,60": 0
+}{
+ "20,70": 0
+}{
+ "20,80": 0
+}{
+ "20,90": 1
+}{
+ "20,100": 1
+}{
+ "20,110": 1
+}{
+ "20,120": 1
+}{
+ "20,130": 1
+}{
+ "20,140": 1
+}{
+ "30,10": 0
+}{
+ "30,20": 0
+}{
+ "30,30": 0
+}{
+ "30,40": 0
+}{
+ "30,50": 0
+}{
+ "30,60": 0
+}{
+ "30,70": 0
+}{
+ "30,80": 1
+}{
+ "30,90": 1
+}{
+ "30,100": 1
+}{
+ "30,110": 1
+}{
+ "30,120": 1
+}{
+ "30,130": 1
+}{
+ "30,140": 1
+}{
+ "40,10": 0
+}{
+ "40,20": 0
+}{
+ "40,30": 0
+}{
+ "40,40": 0
+}{
+ "40,50": 0
+}{
+ "40,60": 0
+}{
+ "40,70": 1
+}{
+ "40,80": 1
+}{
+ "40,90": 1
+}{
+ "40,100": 1
+}{
+ "40,110": 1
+}{
+ "40,120": 1
+}{
+ "40,130": 1
+}{
+ "40,140": 1
+}{
+ "50,10": 0
+}{
+ "50,20": 0
+}{
+ "50,30": 0
+}{
+ "50,40": 0
+}{
+ "50,50": 0
+}{
+ "50,60": 1
+}{
+ "50,70": 1
+}{
+ "50,80": 1
+}{
+ "50,90": 1
+}{
+ "50,100": 1
+}{
+ "50,110": 1
+}{
+ "50,120": 1
+}{
+ "50,130": 1
+}{
+ "50,140": 1
+}{
+ "60,10": 0
+}{
+ "60,20": 0
+}{
+ "60,30": 0
+}{
+ "60,40": 0
+}{
+ "60,50": 1
+}{
+ "60,60": 1
+}{
+ "60,70": 1
+}{
+ "60,80": 1
+}{
+ "60,90": 1
+}{
+ "60,100": 1
+}{
+ "60,110": 1
+}{
+ "60,120": 1
+}{
+ "60,130": 1
+}{
+ "60,140": 1
+}{
+ "70,10": 0
+}{
+ "70,20": 0
+}{
+ "70,30": 0
+}{
+ "70,40": 1
+}{
+ "70,50": 1
+}{
+ "70,60": 1
+}{
+ "70,70": 1
+}{
+ "70,80": 1
+}{
+ "70,90": 1
+}{
+ "70,100": 1
+}{
+ "70,110": 1
+}{
+ "70,120": 1
+}{
+ "70,130": 1
+}{
+ "70,140": 1
+}{
+ "80,10": 0
+}{
+ "80,20": 0
+}{
+ "80,30": 1
+}{
+ "80,40": 1
+}{
+ "80,50": 1
+}{
+ "80,60": 1
+}{
+ "80,70": 1
+}{
+ "80,80": 1
+}{
+ "80,90": 1
+}{
+ "80,100": 1
+}{
+ "80,110": 1
+}{
+ "80,120": 1
+}{
+ "80,130": 1
+}{
+ "80,140": 1
+}{
+ "90,10": 0
+}{
+ "90,20": 1
+}{
+ "90,30": 1
+}{
+ "90,40": 1
+}{
+ "90,50": 1
+}{
+ "90,60": 1
+}{
+ "90,70": 1
+}{
+ "90,80": 1
+}{
+ "90,90": 1
+}{
+ "90,100": 1
+}{
+ "90,110": 1
+}{
+ "90,120": 1
+}{
+ "90,130": 1
+}{
+ "90,140": 1
+}{
+ "100,10": 1
+}{
+ "100,20": 1
+}{
+ "100,30": 1
+}{
+ "100,40": 1
+}{
+ "100,50": 1
+}{
+ "100,60": 1
+}{
+ "100,70": 1
+}{
+ "100,80": 1
+}{
+ "100,90": 1
+}{
+ "100,100": 1
+}{
+ "100,110": 1
+}{
+ "100,120": 1
+}{
+ "100,130": 1
+}{
+ "100,140": 1
+}{
+ "110,10": 1
+}{
+ "110,20": 1
+}{
+ "110,30": 1
+}{
+ "110,40": 1
+}{
+ "110,50": 1
+}{
+ "110,60": 1
+}{
+ "110,70": 1
+}{
+ "110,80": 1
+}{
+ "110,90": 1
+}{
+ "110,100": 1
+}{
+ "110,110": 1
+}{
+ "110,120": 1
+}{
+ "110,130": 1
+}{
+ "110,140": 1
+}{
+ "120,10": 1
+}{
+ "120,20": 1
+}{
+ "120,30": 1
+}{
+ "120,40": 1
+}{
+ "120,50": 1
+}{
+ "120,60": 1
+}{
+ "120,70": 1
+}{
+ "120,80": 1
+}{
+ "120,90": 1
+}{
+ "120,100": 1
+}{
+ "120,110": 1
+}{
+ "120,120": 1
+}{
+ "120,130": 1
+}{
+ "120,140": 1
+}{
+ "130,10": 1
+}{
+ "130,20": 1
+}{
+ "130,30": 1
+}{
+ "130,40": 1
+}{
+ "130,50": 1
+}{
+ "130,60": 1
+}{
+ "130,70": 1
+}{
+ "130,80": 1
+}{
+ "130,90": 1
+}{
+ "130,100": 1
+}{
+ "130,110": 1
+}{
+ "130,120": 1
+}{
+ "130,130": 1
+}{
+ "130,140": 1
+}{
+ "140,10": 1
+}{
+ "140,20": 1
+}{
+ "140,30": 1
+}{
+ "140,40": 1
+}{
+ "140,50": 1
+}{
+ "140,60": 1
+}{
+ "140,70": 1
+}{
+ "140,80": 1
+}{
+ "140,90": 1
+}{
+ "140,100": 1
+}{
+ "140,110": 1
+}{
+ "140,120": 1
+}{
+ "140,130": 1
+}{
+ "140,140": 1
+}{
+ "150,10": 1
+}{
+ "150,20": 1
+}{
+ "150,30": 1
+}{
+ "150,40": 1
+}{
+ "150,50": 1
+}{
+ "150,60": 1
+}{
+ "150,70": 1
+}{
+ "150,80": 1
+}{
+ "150,90": 1
+}{
+ "150,100": 1
+}{
+ "150,110": 1
+}{
+ "150,120": 1
+}{
+ "150,130": 1
+}{
+ "150,140": 1
+}
